WIP: Linux port  - Fixing various warnings/errors when compiling with GCC & Clang
Former-commit-id: dcd15a9ad79efdac12fa3ef8eae9b31926083211

+++ Source/BansheeEditor/SceneView/BsScenePicking.cpp
@@ -104,7 +104,7 @@ namespace bs
 				continue;
 
 			bool found = false;
-			for (int i = 0; i < ignoreRenderables.size(); i++)
+			for (UINT32 i = 0; i < (UINT32)ignoreRenderables.size(); i++)
 			{
 				if (ignoreRenderables[i] == so)
 				{
